Rewrite of the LoadData class. This version is almost a drop in replacement. The main difference is that it allows loading the data with multiple threads in parallel. But it also conforms strictly to the TPC-C specs with regards to the content. The old loader did not produce NULL values for undelivered orders.

+++ run/sql.common/tableCreates.sql
@@ -108,7 +108,7 @@ create table bmsql_item (
 create table bmsql_stock (
   s_w_id       integer       not null,
   s_i_id       integer       not null,
-  s_quantity   decimal(4,0),
+  s_quantity   integer,
   s_ytd        decimal(8,2),
   s_order_cnt  integer,
   s_remote_cnt integer,
